+++ src/components/TodoList/TodoList.module.css
@@ -1,0 +1,12 @@
+.todos {
+    margin-top: 3rem;
+    display: flex;
+    flex-wrap: wrap;
+    gap: 1.5rem;
+    flex-direction: row;
+    justify-content: center;
+}
+.empty__message{
+    font-size: 2.4rem;
+    margin-top: 3rem;
+}
