+++ src/components/index.js
@@ -4,3 +4,4 @@ export * from "./TodoItem/TodoItem";
 export * from "./Button/Button";
 export * from "./Fields/InputField/InputField";
 export * from "./TabContainer/TabContainer";
+export * from "./TodoList/TodoList";
