+++ src/pages/TodoPage/TodoPage.jsx
@@ -2,17 +2,14 @@ import { useDispatch, useSelector } from "react-redux";
 import { useEffect, useState } from "react";
 import { addTodo, changeTask, cancelEdition } from "../../features/todos.slice";
 import styles from "./TodoPage.module.css";
-import { TodoItem, InputField, Button, TabContainer } from "../../components";
+import { InputField, Button, TabContainer, TodoList } from "../../components";
 
 export function TodoPage() {
-  const { todoForEdit, todoList, filterValue } = useSelector(
-    (state) => state.todos
-  );
+  const { todoForEdit, todoList } = useSelector((state) => state.todos);
   const { username } = useSelector((state) => state.username);
   const dispatch = useDispatch();
   const [taskName, setTaskName] = useState("");
   const [isDisabled, setIsDisabled] = useState(true);
-  const [list, setList] = useState(todoList);
 
   useEffect(() => {
     if (todoForEdit) {
@@ -21,20 +18,6 @@ export function TodoPage() {
       setTaskName("");
     }
   }, [todoForEdit]);
-
-  useEffect(() => {
-    if (filterValue === "active") {
-      filterTodoList(true);
-    } else if (filterValue === "inactive") {
-      filterTodoList(false);
-    } else {
-      setList(todoList);
-    }
-  }, [filterValue, todoList]);
-
-  function filterTodoList(isActive) {
-    setList(todoList.filter((item) => item.active === isActive));
-  }
 
   function reset() {
     setIsDisabled(true);
@@ -117,17 +100,7 @@ export function TodoPage() {
           {todoList.length < 2 ? "task" : "tasks"}
         </h2>
         <TabContainer />
-        <div className={styles.todos__container}>
-          {list.length > 0 ? (
-            <ul className={styles.todos}>
-              {list.map((todo) => (
-                <TodoItem key={todo.id} task={todo} />
-              ))}
-            </ul>
-          ) : (
-            <p className={styles.empty__message}>Your list is empty!!</p>
-          )}
-        </div>
+        <TodoList />
       </div>
     </main>
   );
