+++ src/pages/TodoPage/TodoPage.module.css
@@ -14,7 +14,7 @@
     margin-bottom: 2rem;
     display: flex;
     gap: 2rem;
-    align-items: start;
+    align-items: flex-start;
     justify-content: center;
     background-color: var(--primary);
     border-top-left-radius: 3rem;
@@ -71,16 +71,4 @@ button:hover {
 .user__title {
     font-size: 1.8rem;
     margin-bottom: 1rem;
-}
-.todos {
-    margin-top: 3rem;
-    display: flex;
-    flex-wrap: wrap;
-    gap: 1.5rem;
-    flex-direction: row;
-    justify-content: center;
-}
-.empty__message{
-    font-size: 2.4rem;
-    margin-top: 3rem;
 }
